bug: add validation when deleting a contractor

+++ src/main/java/com/iws_manager/iws_manager_api/services/impl/ContractorServiceImpl.java
@@ -3,12 +3,15 @@ package com.iws_manager.iws_manager_api.services.impl;
 import com.iws_manager.iws_manager_api.models.Contractor;
 import com.iws_manager.iws_manager_api.repositories.ContractorRepository;
 import com.iws_manager.iws_manager_api.services.interfaces.ContractorService;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 @Transactional
@@ -69,6 +72,10 @@ public class ContractorServiceImpl implements ContractorService {
     public void delete(Long id) {
         if (id==null){
             throw new IllegalArgumentException("ID cannot be null");
+        }
+
+        if (!contractorRepository.existsById(id)) {  
+            throw new EntityNotFoundException("Contractor not found with id: " + id);
         }
         contractorRepository.deleteById(id);
     }
